# docs/docs/api_and_sdks/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Philter API",
    "version": "4.0.0-SNAPSHOT"
  },
  "servers": [
    {
      "description": "Philter API",
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/reidentify": {
      "post": {
        "tags": [
          "Re-identification"
        ],
        "summary": "Re-identify redacted values.",
        "description": "Reverses one or more values that were redacted via the CRYPTO_REPLACE or FPE_ENCRYPT_REPLACE strategy. The caller must supply a stated reason, which is recorded in the audit log alongside the actor, timestamp, and count of reversals. A user may re-identify their own values; an admin may re-identify any user's values by supplying that user's email via the owner parameter. CRYPTO_REPLACE requires policyName so the key can be resolved from the stored policy. FPE_ENCRYPT_REPLACE uses the user's per-account FPE key; if the policy specified a custom FPE configuration, also pass policyName so its key and tweak are used instead. Requires the `reidentify` scope.",
        "operationId": "reidentify",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReidentifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReidentifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid request fields.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope, disabled API, or access denied."
          },
          "404": {
            "description": "Named policy not found, or owner not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "406": {
            "description": "No representation matches Accept."
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReidentifyRequest": {
        "type": "object",
        "properties": {
          "policyName": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "strategy": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ReidentifyResponse": {
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReidentifyResult"
            }
          }
        }
      },
      "ReidentifyResult": {
        "properties": {
          "decrypted": {
            "type": "string"
          },
          "encrypted": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "description": "A Philter API key, sent as `Authorization: Bearer <api key>`. Create one on the API Keys page of the dashboard.",
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ],
...
}